remove scripts dir
