--- css/oiplayer.css
@@ -1,1 +1,0 @@
-.oiplayer>.oipcontrols{position:absolute;width:100%;margin:0;z-index:19;color:var(--oiplayer-color);background-color:var(--oiplayer-background-color);-webkit-user-select:none;-moz-user-select:none;-ms-user-select:none;user-select:none}.fullscreen .oiplayer>.oipcontrols{z-index:1003;margin:0 auto}.oiplayer>.oipcontrols .controls{margin:0;padding:0 .25rem;list-style:none;display:flex;align-items:center;justify-content:center;overflow:hidden;width:100%;position:relative}.oiplayer>.oipcontrols .controls button{display:flex;align-items:center;justify-content:center;cursor:pointer;border:none;color:var(--oiplayer-color);background-color:rgba(0,0,0,0);width:var(--oiplayer-button-size);height:var(--oiplayer-button-size);padding:0}.oiplayer>.oipcontrols .controls button:hover{color:var(--oiplayer-color-hover)}.oiplayer>.oipcontrols .controls button svg{display:block}.oiplayer>.oipcontrols .controls .play>button[data-button-play=playing] [data-button-play-icon=playing]{display:inline-block}.oiplayer>.oipcontrols .controls .play>button[data-button-play=playing] [data-button-play-icon=paused]{display:none}.oiplayer>.oipcontrols .controls .play>button[data-button-play=paused] [data-button-play-icon=playing]{display:none}.oiplayer>.oipcontrols .controls .play>button[data-button-play=paused] [data-button-play-icon=paused]{display:inline-block;margin-left:2px}.oiplayer>.oipcontrols .controls .time,.oiplayer>.oipcontrols .controls .timeleft{font:400 .85rem/.9 Monaco,"Courier New",monospaced;font-size:calc(var(--oiplayer-font-size) - .3rem);text-align:center;margin:0 .5rem}.oiplayer>.oipcontrols .controls .progress{position:relative;flex-grow:1;height:var(--oiplayer-bar-height)}.oiplayer>.oipcontrols .controls .progress .bar{pointer-events:none;transition:width 125ms;height:var(--oiplayer-bar-height);border-radius:var(--oiplayer-bar-border-radius);position:absolute;left:0;top:0}.oiplayer>.oipcontrols .controls .progress .bar.back{z-index:1;cursor:pointer;pointer-events:all;background-color:var(--oiplayer-bar-background-color);width:100%}.oiplayer>.oipcontrols .controls .progress .bar.loaded{z-index:2;background-color:var(--oiplayer-loaded-background-color);min-width:var(--oiplayer-bar-height)}.oiplayer>.oipcontrols .controls .progress .bar.played{z-index:3;background-color:var(--oiplayer-played-background-color);min-width:var(--oiplayer-bar-height)}.oiplayer>.oipcontrols .controls .progress .bar.push{z-index:4;background-color:rgba(0,0,0,0);min-width:var(--oiplayer-bar-height);display:flex;align-items:center;justify-content:end}.oiplayer>.oipcontrols .controls .progress .bar.push .pos{display:block;width:var(--oiplayer-bar-height);height:var(--oiplayer-bar-height);border-radius:var(--oiplayer-bar-border-radius);background-color:var(--oiplayer-played-position-color);pointer-events:all}.oiplayer>.oipcontrols .controls .progress .bar.push .pos:hover{background-color:var(--oiplayer-color-hover)}.oiplayer>.oipcontrols .controls .progress .bar.push .pos>span{display:none}.oiplayer>.oipcontrols .controls .screen>button>svg{width:24px;height:16px}.oiplayer>.oipcontrols .controls .sound>button{justify-content:flex-end}.oiplayer>.oipcontrols .controls .sound>button[data-button-volume=on] [data-button-volume-icon=on]{display:inline-block;margin-right:.3rem}.oiplayer>.oipcontrols .controls .sound>button[data-button-volume=on] [data-button-volume-icon=muted]{display:none}.oiplayer>.oipcontrols .controls .sound>button[data-button-volume=muted] [data-button-volume-icon=on]{display:none}.oiplayer>.oipcontrols .controls .sound>button[data-button-volume=muted] [data-button-volume-icon=muted]{display:inline-block;margin-right:.5rem}figure.oiplayer{--oiplayer-color: #808080;--oiplayer-color-hover: #606060;--oiplayer-border-radius: 0.2rem;--oiplayer-box-shadow: 0.1rem 0.35rem 0.6rem rgba(0, 0, 0, 0.5);--oiplayer-background-color: #ddd;--oiplayer-background-color-top: rgba(255, 255, 255, 0.9);--oiplayer-button-size: 2rem;--oiplayer-font-size: 1rem;--oiplayer-bar-height: 0.75rem;--oiplayer-bar-border-radius: 0.6rem;--oiplayer-bar-background-color: #ffffff;--oiplayer-loaded-background-color: #cccccc;--oiplayer-played-background-color: #a6a6a6;--oiplayer-played-position-color: #808080;--oiplayer-figure-background-color: rgba(0, 0, 0, 0.85)}figure.oiplayer.dark{--oiplayer-color: whitesmoke;--oiplayer-color-hover: #555;--oiplayer-background-color: #1a1a1a;--oiplayer-background-color-top: rgba(0, 0, 0, 0.85);--oiplayer-bar-background-color: whitesmoke;--oiplayer-loaded-background-color: #cccccc;--oiplayer-played-background-color: #a6a6a6;--oiplayer-player-position-color: #1a1a1a}figure.oiplayer{box-sizing:border-box;-webkit-font-smoothing:antialiased}figure.oiplayer *,figure.oiplayer *::before,figure.oiplayer *::after{box-sizing:inherit}figure.oiplayer a:focus{outline:none;background-color:rgba(0,0,0,0)}figure.oiplayer{position:relative;display:inline-block;margin:0;padding:0;font:400 1rem/1 "Helvetica Neue",Helvetica,Arial,sans-serif;font-size:var(--oiplayer-font-size, 1rem);background-color:var(--oiplayer-figure-background-color);height:var(--oiplayer-height, auto);width:var(--oiplayer-width, auto)}figure.oiplayer.audio .oipcontrols{bottom:0}figure.oiplayer.audio .oipcontrols .screen{display:none}figure.oiplayer.top{overflow:hidden}figure.oiplayer.top .oipcontrols{bottom:0;left:0;background-color:var(--oiplayer-background-color-top);border-radius:var(--oiplayer-border-radius);box-shadow:var(--oiplayer-box-shadow);margin:0 1rem 1rem 1rem;width:calc(100% - 2rem);transition:opacity .35s,transform .35s}figure.oiplayer.top .oipcontrols[data-show=show]{opacity:1;transform:translateY(0)}figure.oiplayer.top .oipcontrols[data-show=hidden]{transition-delay:.75s;opacity:0;transform:translateY(3rem)}figure.oiplayer .preview{position:absolute;top:0;left:0;z-index:9;cursor:pointer}figure.oiplayer .preview[data-preview=hidden]{display:none}figure.oiplayer:fullscreen{display:flex;flex-direction:column;justify-content:space-between;max-width:100%;height:100%;background-color:rgba(0,0,0,0)}figure.oiplayer:fullscreen video{width:100%;height:100%}figure.oiplayer:fullscreen .oipcontrols{position:static;margin:0 auto;max-width:80%}figure.oiplayer:fullscreen .preview,figure.oiplayer:fullscreen figcaption{display:none}.oiplayerinfo{font-size:xx-small;padding:2px;background-color:#efefef;color:#000;z-index:99;border-top:1px solid #bbb;font-family:Monaco,"Courier New",monospaced;overflow:auto;position:absolute;top:0;left:0}.oiplayerinfo.bottom{position:static;margin-top:32px}#clientcaps{behavior:url(#default#clientCaps);display:none}body.oiplayer-example{font:400 1rem/1.4 "Helvetica Neue",Helvetica,Arial,sans-serif;background-color:#efefef;line-height:1.4em}body.oiplayer-example a{color:#c00}body.oiplayer-example a:visited{color:#333}body.oiplayer-example a:hover,body.oiplayer-example a:active{color:#c00;text-decoration:underline}body.oiplayer-example>div.main{width:640px;margin:0 auto;background-color:#fff;padding:1px 1em}body.oiplayer-example>div.main ol{padding-left:20px}body.oiplayer-example .license{display:block;font:400 1rem/1 "Helvetica Neue",Helvetica,Arial,sans-serif}
