--- js/oiplayer.js
@@ -1,1 +1,0 @@
-!function(t,e){"object"==typeof exports&&"object"==typeof module?module.exports=e():"function"==typeof define&&define.amd?define([],e):"object"==typeof exports?exports.OIPlayer=e():t.OIPlayer=e()}(this,()=>(()=>{"use strict";var t={d:(e,i)=>{for(var s in i)t.o(i,s)&&!t.o(e,s)&&Object.defineProperty(e,s,{enumerable:!0,get:i[s]})},o:(t,e)=>Object.prototype.hasOwnProperty.call(t,e)},e={};t.d(e,{default:()=>a});class i{constructor(t,e,i){this.media=t,this.oiplayer=e,this.config=i,this.myname="super",this.init()}mute(){}play(){}pause(){}seek(t){}info(){}volume(t){}init(){this.state="init",this.updatedMetadata=0,this.bufferedSeconds=0,this.poster=this.media.getAttribute("poster")}get height(){const t="audio"===this.type?48:288;return parseInt(this.media.getAttribute("height"))||t}get width(){return parseInt(this.media.getAttribute("width"))||512}get length(){return-1}get position(){return 0}get type(){return this.media.tagName.toLowerCase()}}class s extends i{constructor(t,e,i){super(t,e,i),this.myname="mediaplayer"}init(){super.init(),this.eventHandlers()}eventHandlers(){this.media.addEventListener("loadedmetadata",()=>{this.metadataLoaded()}),this.media.addEventListener("timeupdate",()=>{this.updatedMetadata<1&&this.metadataLoaded()}),this.media.addEventListener("progress",()=>{if(this.media.duration>0)for(let t=0;t<this.media.buffered.length;t++)if(this.media.buffered.start(this.media.buffered.length-1-t)<this.media.currentTime){this.bufferedSeconds=this.media.buffered.end(this.media.buffered.length-1-t);break}}),this.media.addEventListener("playing",()=>this.state="playing"),this.media.addEventListener("pause",()=>this.state="paused"),this.media.addEventListener("ended",()=>this.state="ended")}metadataLoaded(){const t=this.media.duration;if(!Number.isFinite(t))return;const e={duration:t,height:this.media.videoHeight?this.media.videoHeight:-1,width:this.media.videoWidth?this.media.videoWidth:-1};this.oiplayer.updateMetadata(e),this.updatedMetadata+=1}static canPlay(t){const e=t.querySelectorAll("source");let i={canplay:"",proposal:"media"};if(!e.length){const e=t.getAttribute("src");if(e)return i={...i,mimetype:"",url:e},i}return e.forEach(e=>{switch(t.canPlayType(e.type)){case"probably":"probably"!==i.canplay&&(i={...i,canplay:"probably",mimetype:e.type,url:e.src});break;case"maybe":"probably"!==i.canplay&&(i={...i,canplay:"maybe",mimetype:e.type,url:e.src})}}),i}play(){this.media.paused||this.media.ended?(this.media.play(),this.state="playing"):(this.media.pause(),this.state="paused")}pause=()=>this.media.pause();seek=t=>{this.media.fastSeek?this.media.fastSeek(t):this.media.currentTime=t};volume=t=>{this.media.muted=!this.media.muted,this.oiplayer.muted=this.media.muted};get length(){return this.media.duration}get position(){return this.media.currentTime}}const a=class{constructor(t,e){t?(this.media=t,this.id=t.id||"id"+Math.random().toString(16).slice(2),this.config={server:"http://www.openimages.eu",controls:"top",...e},this.init()):console.error("OIPlayer needs at least a video or audio element!")}init(){this.state="init",this.media.controls=!1;const t=document.createElement("figure");t.classList.add("oiplayer"),this.media.replaceWith(t),t.appendChild(this.media),this.figure=t;const e=s.canPlay(this.media),i=this.config;this.config={...i,...e},this.player=new s(this.media,this,this.config),this.oipAttributes(),this.makeUI(),this.handlers(),this.events()}makeUI(){const t=this.makeControlsHtml(),e=document.createElement("div");e.innerHTML=t,e.classList.add("oipcontrols"),this.figure.appendChild(e),this.oipcontrols=e;const i=this.makePreviewHtml();i&&(this.figure.appendChild(i),this.previewScreen=i),this.figure.classList.add(this.type),this.controlsDark&&this.figure.classList.add("dark"),this.controlsTop&&this.figure.classList.add("top"),this.setHeightWidth()}setHeightWidth(){this.figure.style.setProperty("--oiplayer-height",`${this.height}px`),this.figure.style.setProperty("--oiplayer-width",`${this.width}px`)}handlers(){this.buttonPlay=this.oipcontrols.querySelector("[data-button-play]"),this.buttonScreen=this.oipcontrols.querySelector("[data-button-screen]"),this.buttonVolume=this.oipcontrols.querySelector("[data-button-volume]"),this.progressPush=this.oipcontrols.querySelector("[data-progress='push']"),this.progressPlayed=this.oipcontrols.querySelector("[data-progress='played']"),this.progressLoaded=this.oipcontrols.querySelector("[data-progress='loaded']"),this.progressBack=this.oipcontrols.querySelector("[data-progress='back']"),this.time=this.oipcontrols.querySelector("[data-time]"),this.timeleft=this.oipcontrols.querySelector("[data-timeleft]"),this.previewScreen=this.figure.querySelector("[data-preview]")}events(){this.buttonPlay.addEventListener("click",()=>this.play()),this.buttonScreen.addEventListener("click",()=>this.fullscreen()),this.buttonVolume.addEventListener("click",()=>this.volume()),this.progressBack.addEventListener("click",t=>this.scrub(t)),this.controlsTop&&(this.figure.addEventListener("mouseover",()=>this.showControls(!0)),this.figure.addEventListener("mouseout",()=>this.showControls(!1)),this.showControls(!0)),this.previewScreen?.addEventListener("click",()=>this.play())}follow(){const t=()=>{const e=this.mediaDuration;e?(this.updateProgress(e,this.player.position),this.updateTime(e,this.player.position),this.updatePlayButton(this.player.state),this.updateLoaded(e),"playing"===this.player.state&&requestAnimationFrame(t)):console.log("no duration",e)};requestAnimationFrame(t),this.hidePreview(),this.changedState(this.player.state)}scrub(t){let e=this.mediaDuration;if(!e)return void console.log("no duration",e);const i=this.progressBack.getBoundingClientRect(),s=(t.clientX-i.left)/this.progressBack.offsetWidth;this.player.seek(s*e),this.follow()}oipAttributes(){const t=this._extraAttributes(this.media);for(let e=0;e<t.length;e++){const i=t[e];"duration"===i.name?this.duration=Number(i.value):"id"===i.name?this.id=`id-${i.value}`:"start"===i.name&&(this.start=Number(i.value))}}updateMetadata(t){const{duration:e,height:i,width:s}=t;this.updateTime(e,this.player.position),this.updateLoaded(e)}updateTime(t,e=0){this.time.innerText=this._totime(e),this.timeleft.innerText=this._totime(t-e)}updatePlayButton(t){"playing"===t?this.buttonPlay.setAttribute("data-button-play","playing"):this.buttonPlay.setAttribute("data-button-play","paused")}updateProgress(t,e=0){const i=Math.round(e/t*1e4)/100;this.progressPlayed.style.width=`${i}%`,this.progressPush.style.width=`${i}%`}updateLoaded(t){const e=this.player.bufferedSeconds;if(e<1)return;const i=Math.round(e/t*1e4)/100;this.progressLoaded.style.width=`${i}%`}play(){this.player.play(),this.follow()}fullscreen(){document?.fullscreenEnabled||console.log("no fullscreen"),null!==document.fullscreenElement?document.exitFullscreen():this.figure.requestFullscreen()}volume(){this.player.volume(),this.buttonVolume.setAttribute("data-button-volume",this.muted?"muted":"on")}showControls(t){t?this.oipcontrols.setAttribute("data-show","show"):"init"!==this.player.state&&this.oipcontrols.setAttribute("data-show","hidden")}makeControlsHtml(){const t=this.mediaDuration?this.mediaDuration:0;return`<ul class="controls" aria-label="Media controls">\n        <li class="play">\n          <button data-button-play="paused" aria-label="Play/Pause">\n            <span data-button-play-icon="playing"><svg width="15" height="18" viewBox="0 0 15 18" fill="none" xmlns="http://www.w3.org/2000/svg"><rect width="5" height="18" fill="currentColor"/><rect x="10" width="5" height="18" fill="currentColor"/></svg></span>\n            <span data-button-play-icon="paused"><svg width="17" height="19" viewBox="0 0 17 19" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M15.5 8.43036C16.1667 8.81526 16.1667 9.77751 15.5 10.1624L2 17.9566C1.33333 18.3415 0.5 17.8604 0.5 17.0906V1.50216C0.5 0.732357 1.33333 0.251233 2 0.636133L15.5 8.43036Z" fill="currentColor" stroke="currentColor"/></svg></span>\n          </button>\n        </li>\n        <li class="time">\n          <div data-time="" aria-label="Time" aria-live="polite">00:00</div>\n        </li>\n        <li class="progress">\n          <div class="bar push" data-progress="push">\n            <button class="pos" aria-label="Position"><span>Position</span></button>\n          </div>\n          <div data-progress="played" class="bar played" aria-label="Played"></div>\n          <div data-progress="loaded" class="bar loaded" aria-label="Loaded"></div>\n          <div data-progress="back" class="bar back"></div>\n        </li>\n        <li class="timeleft">\n          <div data-timeleft="" aria-label="Time left" aria-live="polite">${this._totime(t)}</div>\n        </li>\n        <li class="screen">\n          <button data-button-screen aria-label="Fullscreen"><svg width="19" height="14" viewBox="0 0 19 14" fill="none" xmlns="http://www.w3.org/2000/svg"><rect x="1" y="1" width="17" height="12" stroke="currentColor" stroke-width="2"/></svg></button>\n        </li>\n        <li class="sound">\n          <button data-button-volume="on" aria-label="Mute/Unmute">\n            <span data-button-volume-icon="on"><svg width="21" height="17" viewBox="0 0 21 17" fill="none" xmlns="http://www.w3.org/2000/svg">\n<circle cx="11.5" cy="8.073" r="1" fill="currentColor" stroke="currentColor"/>\n<path d="M0.5 10.873V5.273H3.5L8.5 1.073V15.073L3.5 10.873H0.5Z" fill="currentColor" stroke="currentColor"/>\n<path d="M15.5 1.073C15.5 1.073 19.5 3.323 19.5 8.073C19.5 12.823 15.5 15.073 15.5 15.073" stroke="currentColor" stroke-width="1.7" stroke-linecap="round"/>\n<path d="M13.5 5.073C13.5 5.073 15.5 6.073 15.5 8.073C15.5 10.073 13.5 11.073 13.5 11.073" stroke="currentColor" stroke-width="1.7" stroke-linecap="round"/>\n</svg></span>\n            <span data-button-volume-icon="muted"><svg width="18" height="17" viewBox="0 0 18 17" fill="none" xmlns="http://www.w3.org/2000/svg"><circle cx="15.5" cy="8.073" r="2" fill="currentColor"/><path d="M0.5 10.873V5.273H3.5L8.5 1.073V15.073L3.5 10.873H0.5Z" fill="currentColor" stroke="currentColor"/></svg></span>\n          </button>\n        </li>\n      </ul>`}makePreviewHtml(){const t=this.media.querySelector("img");if(t)return t.setAttribute("data-preview","shown"),t.classList.add("preview"),t;const e=this.media.getAttribute("poster");if(!e)return;const i=document.createElement("img");return i.src=e,i.width=this.width,i.height=this.height,i.setAttribute("data-preview","shown"),i.classList.add("preview"),i}hidePreview(){"video"===this.player.type&&"shown"===this.previewScreen?.getAttribute("data-preview")&&this.previewScreen.setAttribute("data-preview","hidden")}changedState(t){"init"===this.state&&"playing"===t&&window.dispatchEvent(new CustomEvent("oiplayerplay",{detail:{id:this.id,start:this.player.position,state:t}})),this.state=t}_extraAttributes(t){const e=t.getAttribute("class"),i=[];if(e){const t=e.split(" ");for(let e=0;e<t.length;e++){const s=t[e];if(s.indexOf("oip_ea")>-1){const t=s.substring(7),e=t.substring(0,t.indexOf("_")),a=t.substring(t.indexOf("_")+1);i.push({name:e,value:a})}}}return i}_totime(t){function e(t){const e=parseInt(t,10);return e<10?"0"+e:e}return t<0&&(t=0),function(t){const i=Math.floor(t/3600);let s=Math.floor(t/60);const a=Math.floor(t-60*s);return i>=1?(s-=60*i,i+":"+e(s)+":"+e(a)):e(s)+":"+e(a)}(Math.floor(t))}get controlsDark(){return this.config.controls.indexOf("dark")>-1}get controlsTop(){return!("audio"===this.type&&!this.previewScreen)&&this.config.controls.indexOf("top")>-1}get mediaDuration(){let t=this.player.length;return Number.isFinite(t)||(t=this.duration),t}get height(){let t=this.player.height||288;if(this.controlsTop||(t+=48),"audio"===this.type)if(this.previewScreen){const e=Number(this.previewScreen.getAttribute("height"));t=this.controlsTop?e:e+48}else t=48;return t}get width(){let t=this.player.width||512;return"audio"===this.type&&this.previewScreen&&(t=Number(this.previewScreen?.getAttribute("width"))),t}get type(){return this.media.tagName.toLowerCase()}};return e=e.default})());
